+++ src/view/film-details.js
@@ -93,7 +93,7 @@ export default class FilmDetails extends AbstractSmartComponent {
   }
 
   getTemplate() {
-    return createFilmDetailsTemplate(this._film, this._comments);
+    return createFilmDetailsTemplate(this._film);
   }
 
   setClickHandler(callback) {
