--- src/view/film-list-container.js
@@ -1,8 +1,0 @@
-import AbstractComponent from "./abstract-component";
-
-export default class FilmListContainer extends AbstractComponent {
-
-  getTemplate() {
-    return `<div class="films-list__container"></div>`;
-  }
-}
